Add lock of the internal goose table
Problem:

Two automated builds/people start the migration process at the same
time. Currently the results are relatively unpredictable expecially on
data migration tasks.

Solution:

Before starting the migration we lock the goose version table resulting
in an error if a second lock attempt is made.
This would force the second migration to back off and retry later. It is
important that the process of acquiring the lock results in an error and
does not just wait for the lock to be released (as it seems to be the
case for Mysql, RedShift) as that would still result in undesired
behaviour.

Unfortunately only PG seems to support locking of the nature that is
required and is thus the only system implemented. The lock call in other
dialects result in a NOOP, but provides a framework and commentary.