reattempt to insert version zero row if it failed initially
This can happen in a situation where the database is under heavy load
and the table is successfully created but the insertion of the version
zero row fails. Without this fix, goose will never be able to recover
from this state.

Note that even though the table creation and row insertion happen under
a single transaction, many database systems do not support transactions
for DDL statements, so the table will be created immediately and the
operations will not be atomic.