Extend the SQLDialect interface
**What**
- Add a `getContext` method
- Add a `rollbackCreateVersionTable method

**Why**
The `getContext` method is required to for dialects, such as snowflake, where query execution may need to be controlled via the context.
The `rollbackCreateVersionTable` method is required for dialects, such as snowflake, where multiple DDL statements cannot be executed in a single transaction but multiple statements are required to create the version table.