Implement SnowflakeDIalect interface
**Note**
Snowflake automatically commits each DDL statement. In order to create the version table with a "serial" primary key column, a sequence must first be created, followed by the table with the primary key column defaulted to `seq.nextval`. In order to support this, multi-statement execution must be enabled via the context. Furthermore, in the event of a failure, the table and sequence creation must be rolled back explicitly.