feat: Provider Verbose logging
Currently, `Provider` with `WithVerbose` is not honored by the provider.

Simplest approach was to update `runGo` and `runSQL` to make these functions Pointer Receiver of `Provider` in order to have a provider `Config` reachable.

Another approach would be to overload the `Context` with `ContextWithValue` and set `cfg`, which is dirtier IMHO because it leaves us with an interface casting

Reusing the global `verbose` variable and verbose logging is not a good idea either because this is not a per provider behaviour.